requirements_client.txt
mss
Pillow
pynput
psutil
pystray
requests
cryptography
